requirements.txt
flask==2.3.2
flask-cors==4.0.0
PyPDF2==3.0.1
pinecone-client==3.0.0
requests>=2.32.3
python-dotenv==1.0.0
python-dateutil==2.8.2
pyreadline3==3.4.1
numpy==1.24.4
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.33.3
torch==2.1.2
accelerate==0.27.2
spacy==3.8.7
concurrent-log-handler==0.9.24
psutil==5.9.8